renaming some modules
